cistarlark: move starlark into a separate package
